hashing, edit/delete video bug fixed
# 8.14
Password hashing bug fixed.
Only video's onwer can edit/delete video

+++ src/controllers/videoController.js
@@ -17,7 +17,7 @@ export const watch = async (req, res) => {
     const video = await Video.findById(id).populate("owner");
     // console.log(video);
     if (!video) {
-        return res.render("404", { pageTitle: "Video Not Found" });
+        return res.status(404).render("404", { pageTitle: "Video Not Found" });
     }
     return res.render("watch", { pageTitle: video.title, video });
 };
@@ -25,10 +25,16 @@ export const watch = async (req, res) => {
 
 export const getEdit = async (req, res) => {
     const { id } = req.params;
+    const {
+        user: { _id }
+    } = req.session;
     const video = await Video.findById(id);
     // console.log(video);
     if (!video) {
         return res.status(404).render("404", { pageTitle: "Video Not Found" });
+    }
+    if (String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
     }
     return res.render("edit", { pageTitle: `Edit: ${video.title}`, video});
 };
@@ -36,10 +42,16 @@ export const getEdit = async (req, res) => {
 
 export const postEdit = async (req, res) => {
     const { id } = req.params;
+    const {
+        user: { _id }
+    } = req.session;
     const { title, description, hashtags } = req.body;
     const video = await Video.exists({ _id: id });
     if (!video) {
         return res.render("404", { pageTitle: "Video Not Found" });
+    }
+    if (String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
     }
     await Video.findByIdAndUpdate(id, {
         title,
@@ -83,6 +95,17 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
     const { id } = req.params;
+    const {
+        user: { _id }
+    } = req.session;
+    const video = await Video.findById(id);
+    if (!video) {
+        return res.status(404).render("404", { pageTitle: "Video Not Found" });
+    }
+    if (String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
+    }
+
     await Video.findByIdAndDelete(id);
     return res.redirect("/");
 };
